+++ rest-api/Core/DTOs/CoinApi/CoinApiAssetHistoryDto.cs
@@ -3,5 +3,5 @@
 public class CoinApiAssetHistoryDto
 {
     public string? PriceUsd { get; set; } = null;
-    public long? Time { get; set; } = null;
+    public decimal? Time { get; set; } = null;
 }
